uses qs to parse and stringify query params, moves defaults to config
