+++ package.json
@@ -61,7 +61,7 @@
     "js-cookie": "^2.1.3",
     "lodash": "^4.17.4",
     "moment": "^2.17.1",
-    "qs": "^6.3.0",
+    "qs": "^6.3.1",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.2",
